⚡️ Feat: 타입별 모달 분기
