+++ src/components/Modal/FollowUserList/FollowUserList.module.scss
@@ -1,0 +1,46 @@
+@import '/src/styles/main.scss';
+
+.user-list {
+  @include column-flexbox($align: start, $justify: start);
+  width: 100%;
+  gap: 25px;
+  overflow-y: scroll;
+  scrollbar-width: none;
+  &::-webkit-scrollbar {
+    display: none;
+  }
+  @include tablet {
+    gap: 20px;
+  }
+  @include mobile {
+    gap: 20px;
+  }
+}
+
+.user-information {
+  @include flexbox;
+  gap: 20px;
+}
+
+.user-image {
+  border-radius: 50%;
+  @include tablet {
+    width: 48px;
+    height: 48px;
+  }
+  @include mobile {
+    width: 48px;
+    height: 48px;
+  }
+}
+
+.user-name {
+  font-size: 18px;
+  font-weight: 500;
+  @include tablet {
+    font-size: 16px;
+  }
+  @include mobile {
+    font-size: 16px;
+  }
+}
