+++ src/components/Modal/ProfileFollowModal/ProfileFollowModal.module.scss
@@ -30,48 +30,3 @@
     padding-right: 8px;
   }
 }
-
-.user-list {
-  @include column-flexbox($align: start, $justify: start);
-  width: 100%;
-  gap: 25px;
-  overflow-y: scroll;
-  scrollbar-width: none;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-  @include tablet {
-    gap: 20px;
-  }
-  @include mobile {
-    gap: 20px;
-  }
-}
-
-.user-information {
-  @include flexbox;
-  gap: 20px;
-}
-
-.user-image {
-  border-radius: 50%;
-  @include tablet {
-    width: 48px;
-    height: 48px;
-  }
-  @include mobile {
-    width: 48px;
-    height: 48px;
-  }
-}
-
-.user-name {
-  font-size: 18px;
-  font-weight: 500;
-  @include tablet {
-    font-size: 16px;
-  }
-  @include mobile {
-    font-size: 16px;
-  }
-}
